+++ final/Foundation4/RunningExercise.cs
@@ -2,7 +2,7 @@ public class Running : Exercise
 {
     private double _distance;
 
-    public Running(string date, int length, double distance) : base(date,length)
+    public Running(string date, double length, double distance) : base(date,length)
     {
         _distance = distance;
     }
